% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnMeans.R
\name{columnMeans}
\alias{columnMeans}
\title{Column Means}
\usage{
columnMeans(data, na.rm = TRUE, ...)
}
\arguments{
\item{data}{A matrix or dataframe.}

\item{na.rm}{Logical. Should missing values (including NaN) be omitted from
the calculations?}

\item{...}{Additional arguments to be passed to \code{\link[base]{colMeans}}.}
}
\description{
Return the column means of a dataframe or matrix as a dataframe.
}
\author{
Nathan Eastwood
}

