% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHexMapData.R
\name{createHexMapData}
\alias{createHexMapData}
\title{Extract the hexagon map data}
\usage{
createHexMapData()
}
\description{
Extract the data from the svg file to produce the local area district map of
the UK.
}
\author{
Nathan Eastwood, Douglas Ashton
}

