% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLeafletData.R
\name{createLeafletData}
\alias{createLeafletData}
\title{Create data for the leaflet map}
\usage{
createLeafletData(data)
}
\arguments{
\item{data}{The statistics to be combined with the shapefile data.}
}
\description{
Merge the statistic data with the map data for the leaflet plot
}
\author{
Nathan Eastwood
}

