% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlotlyData.R
\name{createPlotlyData}
\alias{createPlotlyData}
\title{Create the map data}
\usage{
createPlotlyData(data)
}
\arguments{
\item{data}{The statistics to be combined with the shapefile data.}
}
\description{
Extract the data from the shapefile and merge it with the local authority
district statistics.
}
\author{
Nathan Eastwood
}

