% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employmentProspects.R
\docType{package}
\name{employmentProspects}
\alias{employmentProspects}
\alias{employmentProspects-package}
\title{Future Employment Prospects App}
\description{
A package housing the future employment prospects across Great Britain app
for the Cabinet Office's Economic and Domestic Affairs Secretariat (EDS).
}

