% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{empTime}
\alias{empTime}
\title{Employment Statistics}
\format{A \code{data.frame} with 1895 rows and 4 columns.}
\usage{
data(empTime)
}
\description{
The employment rates for each LAD from 2012 to 2016.
}
\keyword{datasets}
