% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHexMapData.R
\name{getAttr}
\alias{getAttr}
\title{Extract individual hexagon data}
\usage{
getAttr(x)
}
\arguments{
\item{x}{A single element of the list extracted from the SVG.}
}
\description{
For each hexagon, extract the corner points (coordinates), the region ID and
the region name.
}
\author{
Nathan Eastwood, Douglas Ashton
}

