% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hexMapData}
\alias{hexMapData}
\title{UK Hex Map Data}
\format{A \code{data.frame} with 2346 rows and 4 columns.}
\usage{
data(hexMapData)
}
\description{
Coordinates for Local Area Districts (LADs) broken down into equally sized
hexagons.
}
\keyword{datasets}

