% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hexMapJson}
\alias{hexMapJson}
\title{UK Hex Map Data}
\format{A \code{list} of length 391 - one for each LAD.}
\usage{
data(hexMapJson)
}
\description{
The LAD data and coordinates for Local Area Districts (LADs) broken down into
equally sized hexagons. The data are of class
\code{SpatialPolygonsDataFrame}.
}
\keyword{datasets}
