% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPage.R
\name{mapPage}
\alias{mapPage}
\title{Server side functionality for maps pages}
\usage{
mapPage(input, output, session, emp, data, cols, hex = TRUE)
}
\arguments{
\item{input}{Shiny parameter.}

\item{output}{Shiny parameter.}

\item{session}{Shiny parameter.}

\item{emp}{The csv input data.}

\item{data}{The data for the map graphic.}

\item{cols}{The names of the columns to display in the datatable.}

\item{hex}{Logical; whether to produce a hexagonal map (\code{TRUE}) or a
geographical map (\code{FALSE}).}
}
\description{
This function works with the \code{\link{mapPageInput}} UI function. It
provides the server functionality.
}
\author{
Nathan Eastwood
}

