% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPageInput.R, R/twoMapPageInput.R
\name{mapPageInput}
\alias{mapPageInput}
\alias{twoMapPageInput}
\title{Input a page for map graphics}
\usage{
mapPageInput(id)

twoMapPageInput(id)
}
\arguments{
\item{id}{The id for the module}
}
\description{
Input either the hexagonal or geographical maps into a page within the app.

Input both the hexagonal or geographical maps into a page within the app.
}
\author{
Nathan Eastwood
}
