% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPlot.R
\name{mapPlot}
\alias{mapPlot}
\title{Create the map of the UK}
\usage{
mapPlot(data, fill)
}
\arguments{
\item{data}{The dataset produced by \code{\link{createPlotlyData}}}

\item{fill}{The statistic by which the map should be filled}
}
\description{
Generate a map of the UK which interactively fills each Local Authority
District (LAD) based on its statistics.
}
\author{
Nathan Eastwood
}

