% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHexMap.R
\name{plotHexMap}
\alias{plotHexMap}
\title{Plot the UK hexagon map}
\usage{
plotHexMap(data, stat = NULL, plotly = TRUE)
}
\arguments{
\item{data}{Data extracted using the \code{\link{createHexMapData}} and saved
as the \code{link{hexMapData}} dataset.}

\item{stat}{The statistic to fill the hexagons with. See details for more
informtation.}

\item{plotly}{Logical. Return the plot using plotly (\code{TRUE}) or as a
ggplot object.}
}
\description{
Draw each local area district of the UK as a hexagon
}
\details{
The \code{stat} data is not included in \code{link{hexMapData}} dataset. You
will need to join data to this dataset. Typically you can join data together
using the \code{lad15cd} column.
}
\author{
Nathan Eastwood, Douglas Ashton
}

