% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roundUp}
\alias{roundUp}
\title{Round up to nearest number}
\usage{
roundUp(x, to = 10)
}
\arguments{
\item{x}{A vector.}

\item{to}{The nearest digit to round up to.}
}
\description{
Round up to nearest number
}
\author{
Nathan Eastwood
}
