% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEmploymentProspectsApp.R
\name{runEmploymentProspectsApp}
\alias{runEmploymentProspectsApp}
\title{Run the App}
\usage{
runEmploymentProspectsApp()
}
\description{
Run the future employment prospects app
}
\author{
Nathan Eastwood
}
