% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shape}
\alias{shape}
\title{Shapefile data for the UK}
\format{As S4 object of class \code{"sp"}; see \code{\link[sp]{sp}}.}
\source{
\href{http://geoportal.statistics.gov.uk/datasets/8edafbe3276d4b56aec60991cbddda50_4}{ONS Geoportal}
}
\usage{
data(shape)
}
\description{
Ultra Generalised Clipped Shapefile data for the Local Authority Districts
(LAD) in the UK from December 2015.
}
\keyword{datasets}
