% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletMap.R
\name{styleMap}
\alias{styleMap}
\title{Apply a style to a leaflet map}
\usage{
styleMap(map, style = list(background = "transparent"))
}
\arguments{
\item{map}{An object of class \code{leaflet}.}

\item{style}{A list containing elements to apply to a \code{leaflet} object.}
}
\description{
Apply a transparent background to a leaflet map
}
\author{
Nathan Eastwood
}

