% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesPlot.R
\name{timeSeriesPlot}
\alias{timeSeriesPlot}
\title{Time Series Plot}
\usage{
timeSeriesPlot(data, x, y, color)
}
\arguments{
\item{data}{The data to plot}

\item{x}{The column to be plotted on the x-axis}

\item{y}{The column to be plotted on the y-axis}

\item{color}{The column to group the data by}
}
\description{
Create a time series plot of employment statistics
}
\examples{
timeSeriesPlot(empTime, x = "year", y = "val", color = "la_name")

}
\author{
Nathan Eastwood
}
