% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAppNumber.R
\name{updateAppNumber}
\alias{updateAppNumber}
\title{Update the app version number}
\usage{
updateAppNumber()
}
\description{
Update the version number in \code{welcome.md} file to be the same as the
package version number.
}
\author{
Nathan Eastwood
}

