% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zygerria.R
\docType{package}
\name{zygerria}
\alias{zygerria}
\alias{zygerria-package}
\title{Future Employment Prospects App}
\description{
A package housing the future employment prospects across Great Britain app
for the Cabinet Office's Economic and Domestic Affairs Secretariat (EDS).
}
