% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeScatter.R
\name{compositeScatter}
\alias{compositeScatter}
\title{Comparison of Composite Variables}
\usage{
compositeScatter(data)
}
\arguments{
\item{data}{The employment data, see details for more information.}
}
\description{
Create a scatter plot of the composite variables
}
\details{
The \code{data} needs to contain \code{measureA}, \code{measureB} and the
employment rate statistics.
}
\author{
Nathan Eastwood
}

