% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataColumnChoices}
\alias{dataColumnChoices}
\title{UK Employment Statistics Choices}
\format{A \code{data.frame} wth 31 observations and 2 columns.}
\usage{
data(dataColumnChoices)
}
\description{
Choices of UK Employment Statistics available in the data. This data set is
used within the Shiny App and maps the full column names to their short hand
counter parts.
}
\keyword{datasets}

