% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletMap.R
\name{leafMap}
\alias{leafMap}
\title{Generate a leaflet map of the UK}
\usage{
leafMap(mapData, fill = NULL, bounds, hex = FALSE)
}
\arguments{
\item{mapData}{The spatial data.}

\item{fill}{The data which should be used to fill the map polygons. Defaults
to \code{NULL} in the event a statistics hasn't been provided.}

\item{bounds}{The boundaries of the map; generated by \code{\link[sp]{bbox}}.}

\item{hex}{Logical; defaults to \code{FALSE}. Whether to generate a hex map
(\code{TRUE}) or a normal map (\code{FALSE}).}
}
\description{
Generate a leaflet map of the UK with polygons for each LAD region filled
by their statistics.
}
\author{
Nathan Eastwood
}

