% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPage.R, R/twoMapPage.R
\name{mapPage}
\alias{mapPage}
\alias{twoMapPage}
\title{Server side functionality for maps pages}
\usage{
mapPage(input, output, session, emp, data, cols, hex = TRUE)

twoMapPage(input, output, session, emp, geodata, hexdata, cols)
}
\arguments{
\item{input}{Shiny parameter.}

\item{output}{Shiny parameter.}

\item{session}{Shiny parameter.}

\item{emp}{The csv input data.}

\item{data}{The data for the map graphic.}

\item{cols}{The names of the columns to display in the datatable.}

\item{hex}{Logical; whether to produce a hexagonal map (\code{TRUE}) or a
geographical map (\code{FALSE}).}

\item{geodata}{Data for the geographical map}

\item{hexdata}{Data for the hexagonal map}
}
\description{
This function works with the \code{\link{mapPageInput}} UI function. It
provides the server functionality.

This function works with the \code{\link{twoMapPageInput}} UI function. It
provides the server functionality.
}
\author{
Nathan Eastwood
}

